requirements.txt
django-restframework
pillow
django-crispy-forms
crispy-bootstrap4
python-dotenv
django-allauth
django-allauth[socialaccount]
stripe



git init
git add .
git commit -m "first commit"
git branch -M main
git remote add origin https://github.com/Sunil-havanur/MovieTicketBooking.git
git push -u origin main